+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ $LOAD_PATH.unshift(File.expand_path('../lib', __dir__))
 
 require 'simplecov'
 require 'codecov'
+require 'webmock/rspec'
 
 # SimpleCov.minimum_coverage 95
 SimpleCov.start